FEAT: 비밀방을 즉시 입장했을 때에 대해서 비밀번호를 입력받도록 하는 처리 추가

+++ src/pages/Room.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
 import styled from 'styled-components';
@@ -15,6 +15,7 @@ import CamList from '@components/game/CamList';
 import ChatLog from '@components/game/ChatLog';
 import Presenter from '@components/game/Presenter';
 import ScoreBoard from '@components/game/ScoreBoard';
+import EnterPrivateRoomModal from '@components/home/EnterPrivateRoomModal';
 import ContentContainer from '@components/layout/ContentContainer';
 import RoomTemplate from '@components/layout/RoomTemplate';
 
@@ -23,16 +24,21 @@ const Room = () => {
   const navigate = useNavigate();
   const { authorized } = useAuthSocket();
   const { userStreamRef, isMediaSuccess } = useAppSelector((state) => state.userMedia);
-  const { lastEnteredRoom } = useAppSelector((state) => state.gameRoom);
+  const { lastEnteredRoom, broadcastedRooms } = useAppSelector((state) => state.gameRoom);
   const { destroyLocalStream } = useLocalStream();
   const { onAnnounceRoomUpdate, offAnnounceRoomUpdate } = useGameUpdateSocket();
   const { emitUserLeaveRoom, emitJoinRoom, onJoinRoom } = useGameSocket();
   const { onError, offError } = useErrorSocket();
 
+  const [isConfirmedUser, setIsConfirmedUser] = useState<boolean>(false);
+  const [passwordModalVisible, setPasswordModalVisible] = useState<boolean>(false);
+  const parsedId = parseInt(id ?? '', 10);
+
   useEffect(() => {
     if (!isMediaSuccess) {
       navigate('/?roomId=' + id);
     }
+    onAnnounceRoomUpdate();
     return () => {
       offAnnounceRoomUpdate();
       emitUserLeaveRoom();
@@ -47,6 +53,14 @@ const Room = () => {
           hideProgressBar: true,
         });
         navigate('/');
+        return;
+      }
+      if (lastEnteredRoom?.roomId === parsedId) {
+        setIsConfirmedUser(true);
+        return;
+      }
+      if (broadcastedRooms.find((room) => room.roomId === parsedId)?.isSecreteRoom) {
+        setPasswordModalVisible(true);
       }
       return () => {
         destroyLocalStream(userStreamRef);
@@ -54,18 +68,39 @@ const Room = () => {
       };
     }
   }, [userStreamRef]);
+
   useEffect(() => {
-    onAnnounceRoomUpdate();
-    if (authorized && id) {
+    if (authorized && id && isConfirmedUser) {
       const intId = parseInt(id, 10);
       !Number.isNaN(intId) && emitJoinRoom({ roomId: intId, roomPassword: lastEnteredRoom?.password });
       onJoinRoom();
       onError([{ target: 'event', value: 'enter-room', callback: () => navigate('/', { replace: true }) }]);
     }
-  }, [id, authorized]);
+  }, [id, authorized, isConfirmedUser]);
+
+  const handlePasswordModalClose = () => {
+    setPasswordModalVisible(false);
+    // TODO: beforeunload 이벤트 활용 또는 확인창 컴포넌트 구현해 적용
+    // TODO: 바깥을 클릭해도 안닫히게, 모달 주위를 어둡게
+    navigate('/');
+  };
+
+  const passwordValidationSuccessHandler = () => {
+    setIsConfirmedUser(true);
+    setPasswordModalVisible(false);
+  };
 
   return (
     <RoomTemplate>
+      {passwordModalVisible && (
+        <EnterPrivateRoomModal
+          roomId={parsedId}
+          roomTitle={broadcastedRooms.find((room) => room.roomId === parsedId)?.roomTitle}
+          onClose={handlePasswordModalClose}
+          visible={passwordModalVisible}
+          successHandler={passwordValidationSuccessHandler}
+        />
+      )}
       <ContentContainer title="SCORE" lights={true}>
         <ScoreBoard />
       </ContentContainer>
